requirements.txt
requests
SpeechRecognition
pyttsx3
google-generativeai
pyaudio
